+++ backend/src/main/java/com/festago/festival/application/validator/school/ExistsFestivalSchoolDeleteValidator.java
@@ -1,9 +1,9 @@
-package com.festago.festival.application;
+package com.festago.festival.application.validator.school;
 
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.ErrorCode;
 import com.festago.festival.repository.FestivalRepository;
-import com.festago.school.application.SchoolDeleteValidator;
+import com.festago.school.application.validator.SchoolDeleteValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Component
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class FestivalSchoolDeleteValidator implements SchoolDeleteValidator {
+public class ExistsFestivalSchoolDeleteValidator implements SchoolDeleteValidator {
 
     private final FestivalRepository festivalRepository;
