+++ backend/src/main/java/com/festago/school/application/validator/SchoolDeleteValidator.java
@@ -1,4 +1,4 @@
-package com.festago.school.application;
+package com.festago.school.application.validator;
 
 public interface SchoolDeleteValidator {
