+++ backend/src/main/java/com/festago/student/application/StudentSchoolDeleteValidator.java
@@ -2,7 +2,7 @@ package com.festago.student.application;
 
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.ErrorCode;
-import com.festago.school.application.SchoolDeleteValidator;
+import com.festago.school.application.validator.SchoolDeleteValidator;
 import com.festago.student.repository.StudentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
